refactor of request mock util
