+++ test/gb/activity.test.ts
@@ -1,6 +1,6 @@
 import { TokenHandler } from '../../src/gb/auth/tokenHandler';
 import { GameBusClient } from '../../src/gb/gbClient';
-import { mockGameBusRequest } from './gbUtils';
+import { mockRequest } from '../testUtils/requestUtils';
 
 jest.mock('axios');
 
@@ -8,7 +8,7 @@ jest.mock('axios');
 
 describe('with mocked activities get call', () => {
     // Request handler that simply returns empty data for every request
-    const request = mockGameBusRequest(() => {
+    const request = mockRequest(() => {
         return Promise.resolve({
             data: []
         });
